[Silver IV] Title: 피하자, Time: 548 ms, Memory: 5928 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/25379. 피하자/피하자.cc
@@ -14,7 +14,7 @@ int main()
         v[i] %= 2;
     }
 
-    int evenCnt = 0, oddCnt = 0;
+    long long evenCnt = 0, oddCnt = 0;
     int evenLast = 0, oddLast = 0;
 
     for (int i = 0; i < N; i++)
